AssertAlignedMemoryAccessInstruction -> AssertHalfwordAlignmentInstruction
