+++ jolt-core/src/jolt/instruction/lh.rs
@@ -4,8 +4,8 @@ use tracer::{ELFInstruction, MemoryState, RVTraceRow, RegisterState, RV32IM};
 use super::VirtualInstructionSequence;
 use crate::jolt::instruction::{
     add::ADDInstruction, and::ANDInstruction, sll::SLLInstruction, sra::SRAInstruction,
-    virtual_assert_aligned_memory_access::AssertAlignedMemoryAccessInstruction,
-    xor::XORInstruction, JoltInstruction,
+    virtual_assert_halfword_alignment::AssertHalfwordAlignmentInstruction, xor::XORInstruction,
+    JoltInstruction,
 };
 /// Loads a halfword from memory and sign-extends it
 pub struct LHInstruction<const WORD_SIZE: usize>;
@@ -36,9 +36,8 @@ impl<const WORD_SIZE: usize> VirtualInstructionSequence for LHInstruction<WORD_S
             _ => panic!("Unsupported WORD_SIZE: {}", WORD_SIZE),
         };
 
-        let is_aligned =
-            AssertAlignedMemoryAccessInstruction::<WORD_SIZE, 2>(rs1_val, offset_unsigned)
-                .lookup_entry();
+        let is_aligned = AssertHalfwordAlignmentInstruction::<WORD_SIZE>(rs1_val, offset_unsigned)
+            .lookup_entry();
         debug_assert_eq!(is_aligned, 1);
         virtual_trace.push(RVTraceRow {
             instruction: ELFInstruction {
